requirements.txt
argparse==1.4.0
h5py==2.10.0
PyYAML==5.1.2
Pillow==6.2.1
numpy==1.17.3
scipy==1.3.1
torch==1.3.0
torchvision==0.4.1
pytorch-ignite==0.2.1
tensorflow-gpu==2.0.0
tensorboardX==1.9
